Renamed ThoughtsPane to AgentActivityPane (vue files) and usage to reflect final name.
