+++ beaker-vue/src/components/chat-interface/AgentActivityPane.vue
